// src/resource/components/Gauge/Gauge2/type.ts
export interface GaugeTwo {
  title: string
  history: string
  datatag: string
  unit: string
  minValue: number
  maxValue: number
  startAngle: number
  endAngle: number
  splitNumber: number
  axisWidth: number
  axisColor: string
  pointerWidth: number
  pointerColor: string
  pointerLength: string
  pointerOffset: string
  pointStyle: string
  tickShow: boolean
  tickFontColor: string
  tickFontSize: number
  tickDistance: number
  titleFontSize: number
  titleOffset: string
  titleColor: string
  detailFontSize: number
  detailFontColor: string
  detailOffset: string
  splitLength: number
  splitColor: string
  splitWidth: number
  tickLength: number
  tickColor: string
  tickWidth: number
}